Split AMGe into AMGe and AMGe_host
In order to prepare the code for CUDA, we create a base class AMGe and move CPU
only code to AMGe_host
