+++ include/mfmg/amge.hpp
@@ -51,111 +51,6 @@ public:
   build_agglomerate_triangulation(unsigned int agglomerate_id) const;
 
   /**
-   * Compute the eigenvalues and the eigenvectors. This functions takes as
-   * inputs:
-   *  - the number of eigenvalues to compute
-   *  - the tolerance (relative accuracy of the Ritz value)
-   *  - the map between the local cells and the global cells
-   *  - a function that computes the local DoFHandler, the local system sparse
-   *    matrix with its sparsity pattern, the local mass matrix with its
-   *    sparsity pattern, and local sparse matrix.
-   *
-   * The function returns the complex eigenvalues, the associated eigenvectors,
-   * and a vector that maps the dof indices from the local problem to the global
-   * problem.
-   */
-  std::tuple<std::vector<std::complex<double>>,
-             std::vector<dealii::Vector<double>>,
-             std::vector<dealii::types::global_dof_index>>
-  compute_local_eigenvectors(
-      unsigned int n_eigenvalues, double tolerance,
-      dealii::Triangulation<dim> const &agglomerate_triangulation,
-      std::map<typename dealii::Triangulation<dim>::active_cell_iterator,
-               typename dealii::DoFHandler<dim>::active_cell_iterator> const
-          &patch_to_global_map,
-      std::function<void(dealii::DoFHandler<dim> &dof_handler,
-                         dealii::SparsityPattern &system_sparsity_pattern,
-                         dealii::SparseMatrix<ScalarType> &,
-                         dealii::SparsityPattern &mass_sparsity_pattern,
-                         dealii::SparseMatrix<ScalarType> &,
-                         dealii::ConstraintMatrix &)> const &evaluate) const;
-
-  /**
-   * Compute the restriction sparse matrix. The rows of the matrix are the
-   * computed eigenvectors. \p dof_indices_maps are used to map the indices in
-   * \p eigenvectors to the global dof indices.
-   */
-  // dealii::TrilinosWrappers::SparseMatrix has a private copy constructor and
-  // no move constructor. Thus, we pass the output by reference instead of
-  // returning it.
-  void compute_restriction_sparse_matrix(
-      std::vector<dealii::Vector<double>> const &eigenvectors,
-      std::vector<std::vector<dealii::types::global_dof_index>> const
-          &dof_indices_map,
-      dealii::TrilinosWrappers::SparseMatrix &restriction_sparse_matrix) const;
-
-  /**
-   * Output the mesh and the agglomerate ids.
-   */
-  void output(std::string const &filename) const;
-
-  /**
-   *  Build the agglomerates and their associated triangulations.
-   */
-  void
-  setup(std::array<unsigned int, dim> const &agglomerate_dim,
-        unsigned int const n_eigenvalues, double const tolerance,
-        std::function<void(dealii::DoFHandler<dim> &dof_handler,
-                           dealii::SparsityPattern &system_sparsity_pattern,
-                           dealii::SparseMatrix<ScalarType> &,
-                           dealii::SparsityPattern &mass_sparsity_pattern,
-                           dealii::SparseMatrix<ScalarType> &,
-                           dealii::ConstraintMatrix &)> const &evaluate) const;
-
-private:
-  /**
-   * This data structure is empty but it is necessary to use WorkStream.
-   */
-  struct ScratchData
-  {
-    // nothing
-  };
-
-  /**
-   * Structure which encapsulates the data that needs to be copied add the end
-   * of Worstream.
-   */
-  struct CopyData
-  {
-    std::vector<dealii::Vector<double>> local_eigenvectors;
-    std::vector<dealii::types::global_dof_index> local_dof_indices_map;
-  };
-
-  /**
-   * This function encapsulates the different functions that work on an
-   * independent set of data.
-   */
-  void local_worker(
-      unsigned int const n_eigenvalues, double const tolerance,
-      std::function<void(dealii::DoFHandler<dim> &dof_handler,
-                         dealii::SparsityPattern &system_sparsity_pattern,
-                         dealii::SparseMatrix<ScalarType> &,
-                         dealii::SparsityPattern &mass_sparsity_pattern,
-                         dealii::SparseMatrix<ScalarType> &,
-                         dealii::ConstraintMatrix &)> const &evaluate,
-      std::vector<unsigned int>::iterator const &agg_id,
-      ScratchData &scratch_data, CopyData &copy_data);
-
-  /**
-   * This function does nothing but is necessary to use WorkStream.
-   */
-  void
-  copy_local_to_global(CopyData const &copy_data,
-                       std::vector<dealii::Vector<double>> &eigenvectors,
-                       std::vector<std::vector<dealii::types::global_dof_index>>
-                           &dof_indices_maps);
-
-  /**
    * Compute the map between the dof indices of the local DoFHandler and the
    * dof indices of the global DoFHandler.
    */
@@ -166,15 +61,11 @@ private:
       dealii::DoFHandler<dim> const &agglomerate_dof_handler) const;
 
   /**
-   * Build the sparsity pattern of the restriction matrix, i.e., the
-   * Epetra_FECrsGraph in Trilinos nomenclature.
+   * Output the mesh and the agglomerate ids.
    */
-  dealii::TrilinosWrappers::SparsityPattern
-  compute_restriction_sparsity_pattern(
-      std::vector<dealii::Vector<double>> const &eigenvectors,
-      std::vector<std::vector<dealii::types::global_dof_index>> const
-          &dof_indices_maps) const;
+  void output(std::string const &filename) const;
 
+protected:
   MPI_Comm _comm;
   dealii::DoFHandler<dim> const &_dof_handler;
 };
