+++ source/CMakeLists.txt
@@ -1,5 +1,6 @@
 SET(MFMG_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/amge.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/amge_host.cc
   )
 
 SET(MFMG_SOURCES ${MFMG_SOURCES} PARENT_SCOPE)
