+++ source/amge_host.cc
@@ -1,0 +1,15 @@
+/*************************************************************************
+ * Copyright (c) 2017 by the mfmg authors                                *
+ * All rights reserved.                                                  *
+ *                                                                       *
+ * This file is part of the mfmg libary. mfmg is distributed under a BSD *
+ * 3-clause license. For the licensing terms see the LICENSE file in the *
+ * top-level directory                                                   *
+ *                                                                       *
+ * SPDX-License-Identifier: BSD-3-Clause                                 *
+ *************************************************************************/
+
+#include <mfmg/amge_host.templates.hpp>
+#include <mfmg/instantiation.hpp>
+
+INSTANTIATE_DIM_SCALARTYPE(TUPLE(AMGe_host))
