+++ tests/test_eigenvectors.cc
@@ -13,7 +13,7 @@
 
 #include "main.cc"
 
-#include <mfmg/amge.hpp>
+#include <mfmg/amge_host.hpp>
 
 #include <deal.II/distributed/tria.h>
 #include <deal.II/dofs/dof_accessor.h>
@@ -27,11 +27,11 @@ namespace tt = boost::test_tools;
 namespace ut = boost::unit_test;
 
 void diagonal_matrices(dealii::DoFHandler<2> &dof_handler,
+                       dealii::ConstraintMatrix &constraints,
                        dealii::SparsityPattern &system_sparsity_pattern,
                        dealii::SparseMatrix<float> &system_matrix,
                        dealii::SparsityPattern &mass_sparsity_pattern,
-                       dealii::SparseMatrix<float> &mass_matrix,
-                       dealii::ConstraintMatrix &constraints)
+                       dealii::SparseMatrix<float> &mass_matrix)
 {
   dealii::FE_Q<2> fe(1);
   dof_handler.distribute_dofs(fe);
@@ -63,7 +63,7 @@ BOOST_AUTO_TEST_CASE(diagonal, *ut::tolerance(1e-12))
   dealii::FE_Q<2> fe(1);
   dealii::DoFHandler<2> dof_handler(triangulation);
   dof_handler.distribute_dofs(fe);
-  mfmg::AMGe<2, float> amge(MPI_COMM_WORLD, dof_handler);
+  mfmg::AMGe_host<2, float> amge(MPI_COMM_WORLD, dof_handler);
 
   unsigned int const n_eigenvalues = 5;
   std::map<typename dealii::Triangulation<2>::active_cell_iterator,
@@ -104,12 +104,11 @@ BOOST_AUTO_TEST_CASE(diagonal, *ut::tolerance(1e-12))
 }
 
 void constrained_diagonal_matrices(
-    dealii::DoFHandler<2> &dof_handler,
+    dealii::DoFHandler<2> &dof_handler, dealii::ConstraintMatrix &constraints,
     dealii::SparsityPattern &system_sparsity_pattern,
     dealii::SparseMatrix<double> &system_matrix,
     dealii::SparsityPattern &mass_sparsity_pattern,
-    dealii::SparseMatrix<double> &mass_matrix,
-    dealii::ConstraintMatrix &constraints)
+    dealii::SparseMatrix<double> &mass_matrix)
 {
   dealii::FE_Q<2> fe(1);
   dof_handler.distribute_dofs(fe);
@@ -143,7 +142,7 @@ BOOST_AUTO_TEST_CASE(diagonal_constraint, *ut::tolerance(1e-12))
   dealii::FE_Q<2> fe(1);
   dealii::DoFHandler<2> dof_handler(triangulation);
   dof_handler.distribute_dofs(fe);
-  mfmg::AMGe<2, double> amge(MPI_COMM_WORLD, dof_handler);
+  mfmg::AMGe_host<2, double> amge(MPI_COMM_WORLD, dof_handler);
 
   unsigned int const n_eigenvalues = 5;
   std::map<typename dealii::Triangulation<2>::active_cell_iterator,
